debugging dragging classifier
rename sdf_utils -> grid_utils
might revert because it's not working

+++ moonshine/src/moonshine/raster_3d.py
@@ -1,13 +1,13 @@
 import numpy as np
 import tensorflow as tf
 
-from link_bot_pycommon.link_bot_sdf_utils import idx_to_point_3d
+from link_bot_pycommon.grid_utils import idx_to_point_3d
 
 
 def raster_3d(state, pixel_indices, res, origin, h, w, c, k, batch_size: int):
     """ output is 1-channel voxel grid, using the max to reduce values contributed to by different points in the input state """
     res = res[0]
-    n_points = tf.cast(state.shape[1] / 3, tf.int32)
+    n_points = tf.cast(state.shape[1] / 3, tf.int64)
     points = tf.reshape(state, [batch_size, n_points, 3])
 
     # Below is a un-vectorized implementation, which is much easier to read and understand
